Manipulating numbers at the binary level.
This assignment will focus on manipulating integer numbers in Java at the binary level. This assignment requires that you understand how bit masking works and how to apply bitwise operators. The application you build will allow two users to play a simple numbers game. We will be implementing this program in a non-traditional manner. The goal is to see how integers can be used to store multiple data values.